Unit Tests Done

+++ ingredients-service/src/main/java/domain/service/IngredientService.java
@@ -5,15 +5,15 @@ import java.util.List;
 import domain.model.Ingredient;
 
 public interface IngredientService {
-	/*
+	
 	public long create(Ingredient ingredient);
 	
 	public void delete(Ingredient ingredient);
 	
 	public void update(Ingredient ingredient);
 	
-	public Ingredient get(Long id);
-	*/
+	public Ingredient get(int id);
+	
 	public List<Ingredient> getAll();
 	
 }
